% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dluhc_2var_time_series.R
\name{dluhc_2var_time_series}
\alias{dluhc_2var_time_series}
\title{Create a 2 line time series graph in a DLUHC style}
\usage{
dluhc_2var_time_series(
  .data,
  datecol,
  ycol,
  groupcol,
  dateformat = "\%Y-\%m-\%d"
)
}
\arguments{
\item{.data}{A dataframe with the data in a long format with one column for the date and one column for the variable value}

\item{datecol}{The column name of the dataframe which contains the date variable}

\item{ycol}{The column name of the dataframe which contains the value}

\item{groupcol}{The column name which contains the grouping variable (such as country, region or type)}
}
\value{
A ggplot2 object
}
\description{
Create a 2 line time series graph in a DLUHC style
}
\examples{
df <- dplyr::filter(gapminder::gapminder,country=="Australia"|country=="China")

dluhc_2var_time_series(.data=df,datecol = year, ycol = lifeExp, groupcol = country, dateformat = "\%Y")
}
