% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dluhc_actual_predict.R
\name{dluhc_actual_predict}
\alias{dluhc_actual_predict}
\title{Create a single line time series graph in a dluhc theme for a line which has a predicted value}
\usage{
dluhc_actual_predict(
  .data,
  xcol,
  ycol,
  cutdate,
  dottedline = TRUE,
  label_names = c("Predicted", "Actual")
)
}
\arguments{
\item{.data}{A dataframe in long format with 2 columns necesarry: Date and value}

\item{xcol}{The column name which contains the date value in a widely used date format}

\item{ycol}{The column name which contains the values}

\item{cutdate}{The date which the predicted values begin from}

\item{dottedline}{A TRUE/FALSE statement to decide if you want a vertical dotted line on the graph to split the prediction and the recorded values}

\item{label_names}{A vector containing the 2 words you want as the label for your lines, the default is c("Predicted", "Actual")}
}
\description{
Create a single line time series graph in a dluhc theme for a line which has a predicted value
}
