% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dluhc_theme.R
\name{dluhc_theme}
\alias{dluhc_theme}
\title{ggplot function to create standard background and axis labels for a DLUHC style graph}
\usage{
dluhc_theme(size = 2)
}
\arguments{
\item{size}{a variable used to edit the text size. This is not the font size, but works as a multiplier for font size, default is 2}
}
\value{
a ggplot object
}
\description{
ggplot function to create standard background and axis labels for a DLUHC style graph
}
\examples{
df <- dplyr::filter(gapminder::gapminder,country == "Austria")
ggplot(df, aes(x = year, y=lifeExp)) + dluhc_theme()
}
