% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dluhc_time_series.R
\name{dluhc_time_series}
\alias{dluhc_time_series}
\title{Add dluhc theme to ggplot chart}
\usage{
dluhc_time_series(.data, xcol, ycol)
}
\description{
This function creates a times series graph on a standard format based on receiving data where the date code is in the format
}
