% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_timeseries.R
\name{forecast_timeseries}
\alias{forecast_timeseries}
\title{Create a single line time series graph in a dluhc theme for a line which has a predicted/forecast value}
\usage{
forecast_timeseries(
  .data,
  datecol,
  ycol,
  cutdate,
  dateformat = "\%Y-\%m-\%d",
  dottedline = TRUE,
  label_names = c("Recorded", "Forecast")
)
}
\arguments{
\item{.data}{A dataframe in long format with 2 columns necesarry: Date and value}

\item{datecol}{The column name which contains the date value in a widely used date format}

\item{ycol}{The column name which contains the values}

\item{cutdate}{The date which the predicted values begin from}

\item{dateformat}{The format which the date is presented in using the standard R date format, see here for more detail https://www.statology.org/r-date-format/}

\item{dottedline}{A TRUE/FALSE statement to decide if you want a vertical dotted line on the graph to split the prediction and the recorded values}

\item{label_names}{A vector containing the 2 words you want as the label for your lines, the default is c("Recorded", "Forecast")}
}
\value{
a ggplot2 object
}
\description{
Create a single line time series graph in a dluhc theme for a line which has a predicted/forecast value
}
\examples{
df <- dluhctheme::OBR_Forecast
forecast_timeseries(df,Date,ycol=GDP,cutdate = "01/10/2022",dateformat = "\%d/\%m/\%Y", label_names = c("Actual","Predicted"))
}
