% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Help_to_Buy.R
\docType{data}
\name{Help_to_Buy}
\alias{Help_to_Buy}
\title{Help to Buy Completion data}
\format{
## `Help_to_Buy`
A data frame with 334 rows and 2 columns:
\describe{
  \item{LA_Code}{Local Authority ONS Code}
  \item{Completions}{The number of Help to Buy completions recorded in the local authority between April 2012 and December 2021}
}
}
\source{
<https://www.gov.uk/government/statistics/help-to-buy-equity-loan-scheme-data-to-31-december-2021>
}
\usage{
Help_to_Buy
}
\description{
A subset of data from the Help to Buy Quarterly statistical release
Q4 2021, produced by DLUHC
}
\keyword{datasets}
