% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LA_map.R
\name{LA_map}
\alias{LA_map}
\title{Create a standard theme map from an sf object}
\usage{
LA_map(
  .data,
  variable,
  legendtitle = NULL,
  LA_col,
  title = NULL,
  subtitle = NULL,
  titlesize = 1,
  map_colours = c("#FFFFFF", "#012169"),
  year = "2021-22",
  countries = "E",
  save = FALSE,
  filepath = NULL
)
}
\arguments{
\item{.data}{An sf object which has data broken down by local authority. This function only works for data in local authority breakdown only.}

\item{variable}{The column in the .data which has the numeric value to be mapped}

\item{legendtitle}{The text to appear as the title for the legend on the map, set to NULL unless specified}

\item{LA_col}{The column which contains the Local Authority code}

\item{title}{The map title, in black bold font. Set to NULL unless specified}

\item{subtitle}{The map subtitile in smaller grey font than the title. Set to NULL unless specified}

\item{titlesize}{A ratio value to increase or decrease the title and subtitle. Initially set to 1}

\item{map_colours}{The colour which you want to represent the low and high values, must be in vector form}

\item{year}{The financial year which the LA codes you are using relate to, given in quotations}

\item{countries}{The countries which you wish to appear on the map. These are E, E+W, GB and UK for England, England+Wales, Great Britain and United Kingdom respectively}

\item{save}{A TRUE/FALSE statement if you want the file to be exported as a png file. If TRUE, the filepath must be defined}

\item{filepath}{If save is TRUE, filepath is the save location of the png output must be in the form of quotations}
}
\value{
a ggplot map, or a png file if save is TRUE
}
\description{
Create a standard theme map from an sf object
}
\examples{
df <- dluhctheme::Help_to_Buy
a <- LA_map(df,variable = Completions,LA_col = LA_Code,year = "2020-21",countries = "E",save = TRUE,filepath = "Image.png",title="Help to Buy is popular near London",legendtitle = "Completions",titlesize=2)
finalise_plot(a,source_name = "Source: Help to Buy Equity Loan Statistics DLUHC",save_filepath = "outputmap.png",footerfontsize = 13,width_pixels = 560,height_pixels = 640,logo_nudge = -0.03)
}
