% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_line_timeseries.R
\name{one_line_timeseries}
\alias{one_line_timeseries}
\title{Create a single line time series graph in a dluhc theme}
\usage{
one_line_timeseries(.data, datecol, ycol, dateformat = "\%Y-\%m-\%d")
}
\arguments{
\item{.data}{A dataframe in long format with 2 columns necessary: Date and value}

\item{datecol}{The column name which contains the date value in a widely used date format}

\item{ycol}{The column name which contains the values}

\item{dateformat}{is the format which the date is presented in using the standard R date format, see here for more detail https://www.statology.org/r-date-format/}
}
\value{
a ggplot object
}
\description{
Create a single line time series graph in a dluhc theme
}
\examples{
df <- dluhctheme::Net_Additions

one_line_timeseries(.data=df,datecol = year, ycol = lifeExp, dateformat = "\%Y")
}
