% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_line_timeseries.R
\name{two_line_timeseries}
\alias{two_line_timeseries}
\title{Create a 2 line time series graph in a DLUHC style}
\usage{
two_line_timeseries(.data, datecol, ycol, groupcol, dateformat = "\%Y-\%m-\%d")
}
\arguments{
\item{.data}{A dataframe with the data in a long format with one column for the date and one column for the variable value}

\item{datecol}{The column name of the dataframe which contains the date variable}

\item{ycol}{The column name of the dataframe which contains the value}

\item{groupcol}{The column name which contains the grouping variable (such as country, region or type)}
}
\value{
A ggplot2 object
}
\description{
Create a 2 line time series graph in a DLUHC style
}
\examples{
df <- dluhctheme::Social_Housing_Sales

two_line_timeseries(.data=df,datecol = year, ycol = value, groupcol = type, dateformat = "\%Y")
}
