% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dluhc_2var_time_series.R
\name{dluhc_2var_time_series}
\alias{dluhc_2var_time_series}
\title{Creating a 2 line time series graph in a DLUHC style}
\usage{
dluhc_2var_time_series(.data, xcol, ycol, groupcol)
}
\arguments{
\item{.data}{A dataframe with the data in a long format with one column for the date (in the format "%Y-%m-%d","%d/%m/%Y","%d/%m/%y","%e/%m/%Y","%d-%b-%Y" or "%d-%b-%y") and one column for the variable value}

\item{xcol}{The column name of the dataframe which contains the date variable}

\item{ycol}{The column name of the dataframe which contains the value}

\item{groupcol}{The column name which contains the grouping variable (such as country, region or type)}
}
\value{
A ggplot2 object
}
\description{
Creating a 2 line time series graph in a DLUHC style
}
