% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dluhc_LA_map.R
\name{dluhc_LA_map}
\alias{dluhc_LA_map}
\title{Create a standard theme map from an sf object}
\usage{
dluhc_LA_map(.data, variable, LA_var, map_colour = "#012169")
}
\arguments{
\item{.data}{An sf object which has data broken down by local authority. This function only works for data in local authority breakdown only.}

\item{variable}{The column in the .data which has the numeric value to be mapped}

\item{LA_var}{The column which contains the Local Authority code}

\item{map_colour}{The colour which you want to represent the high value, the low value will always be white}
}
\description{
Create a standard theme map from an sf object
}
