% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dluhc_time_series.R
\name{dluhc_time_series}
\alias{dluhc_time_series}
\title{Create a single line time series graph in a dluhc theme}
\usage{
dluhc_time_series(.data, xcol, ycol)
}
\arguments{
\item{.data}{A dataframe in long format with 2 columns necesarry: Date and value}

\item{xcol}{The column name which contains the date value in a widely used date format: "%Y-%m-%d","%d/%m/%Y","%d/%m/%y","%e/%m/%Y","%d-%b-%Y","%d-%b-%y"}

\item{ycol}{The column name which contains the values}
}
\description{
Create a single line time series graph in a dluhc theme
}
