% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet_barchart.R
\name{facet_barchart}
\alias{facet_barchart}
\title{Title}
\usage{
facet_barchart(.data, xcol, valuecol, groupcol, textsize = 1)
}
\arguments{
\item{.data}{The dataframe used for the analysis, ideally a 3 column dataframe}

\item{xcol}{The column which one of the grouping variables will be in, this could be numeric or categorical}

\item{valuecol}{the value column which contains a numeric variables}

\item{groupcol}{the column which contains the category by which the graphs will be faceted}

\item{textsize}{The text size as a ratio:2 to the standard dluhc style text size, set to 1 by default}
}
\description{
Title
}
