% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LA_map.R
\name{LA_map}
\alias{LA_map}
\title{Create a standard theme map from an sf object}
\usage{
LA_map(
  .data,
  variable,
  LA_col,
  map_colours = c("#FFFFFF", "#012169"),
  year = 2022,
  countries = "E",
  save = FALSE,
  filepath = NULL
)
}
\arguments{
\item{.data}{An sf object which has data broken down by local authority. This function only works for data in local authority breakdown only.}

\item{variable}{The column in the .data which has the numeric value to be mapped}

\item{LA_col}{The column which contains the Local Authority code}

\item{map_colours}{The colour which you want to represent the low and high values, must be in vector form}

\item{year}{The year which the LA codes you are using relate to}

\item{countries}{The countries which you wish to appear on the map. These are E, E+W, GB and UK for England, England+Wales, Great Britain and United Kingdom respectively}

\item{save}{A TRUE/FALSE statement if you want the file to be exported as a png file. If TRUE, the filepath must be defined}

\item{filepath}{If save is TRUE, filepath is the save location of the png output must be in the form of quotations}
}
\value{
a ggplot map, or a png file if save is TRUE
}
\description{
Create a standard theme map from an sf object
}
\examples{
df <- dluhctheme::Net_Additions_LA
LA_map(df,variable = Net_Additions_per_1000,LA_col = LA_Code,year = 2021,countries = "E",save = TRUE, filepath = "Net_Additions_map.png")
}
